makes both graphs trade hover & click events
